# docs/swagger.json
{
    "schemes": [
        "http",
        "https"
    ],
    "swagger": "2.0",
    "info": {
        "description": "This is the API documentation for Livotech Warehouse Management System API Documentation",
        "title": "Livotech Warehouse Management System API Documentation",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "email": "support@livo.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0"
    },
    "host": "localhost:8040",
    "basePath": "/",
    "paths": {
...
        "/api/qc-onlines": {
            "get": {
                "description": "Retrieve a list of QC Onlines with pagination and search",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "QC Onlines"
                ],
                "summary": "Get QC Onlines",
                "parameters": [
                    {
                        "type": "integer",
                        "default": 1,
                        "description": "Page number",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 10,
                        "description": "Number of QC Onlines per page",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Search term for tracking number",
                        "name": "search",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/utils.SuccessPaginatedResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/models.QCOnlineResponse"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/utils.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/utils.ErrorResponse"
                        }
                    }
                }
            },
            "post": {
                "description": "Create a new QC Online",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "QC Onlines"
                ],
                "summary": "Create QC Online",
                "parameters": [
                    {
                        "description": "QC Online details",
                        "name": "qcOnline",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/controllers.CreateQCOnlineRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/utils.SuccessResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/models.QCOnlineResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/utils.ErrorResponse"
                        }
                    },
                    "409": {
                        "description": "Conflict",
                        "schema": {
                            "$ref": "#/definitions/utils.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/utils.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/qc-onlines/{id}": {
            "get": {
                "description": "Retrieve a single QC Online by ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "QC Onlines"
                ],
                "summary": "Get QC Online",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "QC Online ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/utils.SuccessResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/models.QCOnlineResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/utils.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/utils.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/utils.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "controllers.CreateQCOnlineRequest": {
            "type": "object",
            "required": [
                "details",
                "trackingNumber"
            ],
            "properties": {
                "details": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/controllers.QCOnlineDetailRequest"
                    }
                },
                "trackingNumber": {
                    "type": "string"
                }
            }
        },
...
        "controllers.QCOnlineDetailRequest": {
            "type": "object",
            "required": [
                "boxId",
                "quantity"
            ],
            "properties": {
                "boxId": {
                    "type": "integer"
                },
                "quantity": {
                    "type": "integer",
                    "minimum": 1
                }
            }
        },
...
        "models.OrderDetailResponse": {
            "type": "object",
            "properties": {
                "price": {
                    "type": "integer"
                },
                "product": {
                    "$ref": "#/definitions/models.ProductResponse"
                },
                "productName": {
                    "type": "string"
                },
                "quantity": {
                    "type": "integer"
                },
                "sku": {
                    "type": "string"
                },
                "variant": {
                    "type": "string"
                }
            }
        },
        "models.OrderResponse": {
            "type": "object",
            "properties": {
                "address": {
                    "type": "string"
                },
                "assignedAt": {
                    "type": "string"
                },
                "assignedBy": {
                    "type": "string"
                },
                "buyer": {
                    "type": "string"
                },
                "canceledAt": {
                    "type": "string"
                },
                "canceledBy": {
                    "type": "string"
                },
                "changedAt": {
                    "type": "string"
                },
                "changedBy": {
                    "type": "string"
                },
                "channel": {
                    "type": "string"
                },
                "complained": {
                    "type": "boolean"
                },
                "courier": {
                    "type": "string"
                },
                "createdAt": {
                    "type": "string"
                },
                "details": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.OrderDetailResponse"
                    }
                },
                "duplicatedAt": {
                    "type": "string"
                },
                "duplicatedBy": {
                    "type": "string"
                },
                "eventStatus": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "orderGineeId": {
                    "type": "string"
                },
                "pendingAt": {
                    "type": "string"
                },
                "pendingBy": {
                    "type": "string"
                },
                "pickedAt": {
                    "type": "string"
                },
                "pickedBy": {
                    "type": "string"
                },
                "processingStatus": {
                    "type": "string"
                },
                "sentBefore": {
                    "type": "string"
                },
                "store": {
                    "type": "string"
                },
                "trackingNumber": {
                    "type": "string"
                },
                "updatedAt": {
                    "type": "string"
                }
            }
        },
...
        "models.ProductResponse": {
            "type": "object",
            "properties": {
                "createdAt": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "image": {
                    "type": "string"
                },
                "location": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "sku": {
                    "type": "string"
                },
                "updatedAt": {
                    "type": "string"
                },
                "variant": {
                    "type": "string"
                }
            }
        },
        "models.QCOnlineDetailResponse": {
            "type": "object",
            "properties": {
                "box": {
                    "type": "string"
                },
                "quantity": {
                    "type": "integer"
                }
            }
        },
        "models.QCOnlineResponse": {
            "type": "object",
            "properties": {
                "complained": {
                    "type": "boolean"
                },
                "createdAt": {
                    "type": "string"
                },
                "details": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.QCOnlineDetailResponse"
                    }
                },
                "id": {
                    "type": "integer"
                },
                "order": {
                    "$ref": "#/definitions/models.OrderResponse"
                },
                "qcBy": {
                    "type": "string"
                },
                "trackingNumber": {
                    "type": "string"
                },
                "updatedAt": {
                    "type": "string"
                }
            }
        },
...
        "utils.ErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string"
                },
                "success": {
                    "type": "boolean"
                }
            }
        },
...
        "utils.Pagination": {
            "type": "object",
            "properties": {
                "limit": {
                    "type": "integer"
                },
                "page": {
                    "type": "integer"
                },
                "total": {
                    "type": "integer"
                }
            }
        },
        "utils.SuccessPaginatedResponse": {
            "type": "object",
            "properties": {
                "data": {},
                "message": {
                    "type": "string"
                },
                "pagination": {
                    "$ref": "#/definitions/utils.Pagination"
                },
                "success": {
                    "type": "boolean"
                }
            }
        },
        "utils.SuccessResponse": {
            "type": "object",
            "properties": {
                "data": {},
                "message": {
                    "type": "string"
                },
                "success": {
                    "type": "boolean"
                }
            }
        }
    }
}